\name{addOutliers-methods}
\docType{methods}
\alias{addOutliers-methods}
\alias{addOutliers,AFLP,AFLP.outlier-method}
\alias{addOutliers,AFLP,data.frame-method}
\title{Add outliers to an AFLP object}
\description{
Add outliers to an AFLP object
}
\section{Methods}{
\describe{

\item{data = "AFLP", extraOutliers = "AFLP.outlier"}{Add outliers as an AFLP.outlier object to an AFLP object}

\item{data = "AFLP", extraOutliers = "data.frame"}{Add outliers as a data.frame to an AFLP object}
}}
\keyword{methods}
\keyword{manip}
