\name{borderProbability}
\Rdversion{1.1}
\alias{borderProbability}
\title{
Internal functions of the AFLP package. 
}
\description{
Not intended for external use.
}
\author{
Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul Quataert
}
\seealso{
\code{\link{AFLP-class}}
}
\keyword{internal}

