\docType{class}
\name{AFLP.outlier-class}
\alias{AFLP.outlier-class}
\title{Outliers of an AFLP object}
\description{
  This class is designed to hold the outliers of an AFLP
  object
}
\section{Objects from the Class}{
  Use \code{AFLP.outlier())} to create objects of this
  class.
}
\examples{
AFLP.outlier()
}
\author{
  Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul
  Quataert
}
\seealso{
  \code{\link{AFLP.outlier}}, \code{\link{markers}},
  \code{\link{replicates}}, \code{\link{residuals}},
  \code{\link{specimens}}
}
\keyword{classes}

