\name{AFLP}
\alias{AFLP}
\title{Create an object of the class AFLP}
\usage{
  AFLP(Specimens, Replicates, QC, Fluorescence, model,
    outliers, Borders, Quality)
}
\arguments{
  \item{Specimens}{A data.frame with specimens}

  \item{Replicates}{A data.frame with replicates}

  \item{QC}{A list with QC samples}

  \item{Fluorescence}{A data.Frame with fluorescence data}

  \item{model}{The formula of the model for normalisation}

  \item{outliers}{AFLP outliers}

  \item{Borders}{The borders between present and absent
  markers}

  \item{Quality}{Quality data}
}
\description{
  Create an object of the class AFLP
}

