\docType{methods}
\name{border-method}
\alias{border-method}
\alias{border,AFLP-method}
\title{Extract the cut-levels from an AFLP object}
\description{
  Extracts the cut-levels from an AFLP object
}
\section{Methods}{
  \describe{

  \item{data = "AFLP"}{Extracts the cut-levels from an AFLP
  object} }
}
\keyword{attribute}
\keyword{methods}

