\name{border<--methods}
\docType{methods}
\alias{border<--methods}
\alias{border<-,AFLP-method}
\title{Adds or overwrites cut-levels}
\description{
Adds or overwrites cut-levels in an AFLP object.
}
\section{Methods}{
\describe{

\item{data = "AFLP"}{\code{data} is updated with the new border information}
}}
\keyword{methods}
