\name{borderProbability}
\alias{borderProbability}
\title{Internal functions of the AFLP package.}
\usage{
  borderProbability(Fluor, Borders, Repeated)
}
\description{
  Not intended for external use.
}
\author{
  Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul
  Quataert
}
\seealso{
  \code{\link{AFLP-class}}
}
\keyword{internal}

