\name{fluorescence-methods}
\docType{methods}
\alias{fluorescence-methods}
\alias{fluoresence,AFLP-method}
\title{Extracting fluorescence from an AFLP object}
\description{
 Extracting fluorescence from an AFLP object
}
\section{Methods}{
\describe{

\item{x = "AFLP"}{Extracting fluorescence from \code{x}}
}}
\keyword{methods}
\keyword{attribute}
