\docType{methods}
\name{fluorescence<--methods}
\alias{fluorescence<--methods}
\alias{fluorescence<-,AFLP,data.frame-method}
\title{Adds or overwrites fluorescence data}
\description{
  Adds or overwrites fluorescence data in an AFLP object.
}
\section{Methods}{
  \describe{

  \item{data = "AFLP", value = "data.frame"}{\code{data} is
  updated with the new fluorescence data} }
}
\keyword{methods}

