\name{fluorescence}
\alias{fluorescence}
\title{Extract the fluorescence (raw, normalised and scored) information from an
AFLP object}
\arguments{
  \item{x}{An AFLP object}
}
\value{
  The \code{Fluorescence} slot of \code{x}.
}
\description{
  Extract the fluorescence (raw, normalised and scored)
  information from an AFLP object
}
\examples{
data(Tilia)
 fluorescence(Tilia)
}
\author{
  Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul
  Quataert
}
\seealso{
  \code{\link{AFLP-class}}
}
\keyword{attribute}

