\name{hclust-methods}
\docType{methods}
\alias{hclust-methods}
\alias{hclust,AFLP-method}
\title{Performs hierarchical clustering}
\description{
Performs hierarchical clustering
}
\section{Methods}{
\describe{

\item{d = "AFLP"}{Performs hierarchical clustering on an AFLP object}

\item{d = "ANY"}{Performs hierarchical clustering}
}}
\keyword{methods}
\keyword{classif}
