\name{markers-method}
\docType{methods}
\alias{markers-method}
\alias{markers,AFLP.outlier-method}
\title{Extract the marker information from an object}
\description{
Extract the marker information from an object
}
\section{Methods}{
\describe{

\item{x = "AFLP.outlier"}{Extract the marker information from an AFLP.outlier object}
}}
\keyword{methods}
\keyword{attribute}
