\name{markers}
\alias{markers}
\title{Extract the marker information from an object}
\arguments{
  \item{x}{An AFLP object or an AFLP.outlier object}
}
\value{
  the marker information from an object
}
\description{
  Extract the marker information from an object
}
\examples{
data(Tilia)
 mTilia <- markers(outliers(Tilia))
}
\author{
  Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul
  Quataert
}
\keyword{attributes}

