\name{outliers-methods}
\docType{methods}
\alias{outliers-methods}
\alias{outliers,AFLP-method}
\title{Extracts outliers}
\description{
Extracts outliers
}
\section{Methods}{
\describe{

\item{x = "AFLP"}{Extracts outliers from an AFLP object}
}}
\keyword{methods}
\keyword{attribute}
