\docType{methods}
\name{outliers-methods}
\alias{outliers-methods}
\alias{outliers,AFLP-method}
\title{Extracts outliers}
\description{
  Extracts outliers
}
\section{Methods}{
  \describe{

  \item{x = "AFLP"}{Extracts outliers from an AFLP object}
  }
}
\keyword{attribute}
\keyword{methods}

