\name{outliers}
\alias{outliers}
\title{Extract outlier information from an AFLP object}
\arguments{
  \item{x}{An AFLP object}
}
\value{
  An object of the AFLP.outlier class with the outliers
  from the AFLP object
}
\description{
  Extract outlier information from an AFLP object
}
\examples{
data(Tilia)
	outliers(Tilia)
}
\author{
  Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul
  Quataert
}
\seealso{
  \code{\link{AFLP-class}},
  \code{\link{AFLP.outlier-class}}
}
\keyword{attribute}

