\docType{methods}
\name{princomp-methods}
\alias{princomp-methods}
\alias{princomp,AFLP-method}
\title{Principal component analysis}
\description{
  Principal component analysis
}
\section{Methods}{
  \describe{

  \item{x = "AFLP"}{PCO on a AFLP object}

  \item{x = "ANY"}{princomp on \code{x}} }
}
\keyword{classif}
\keyword{methods}

