\docType{methods}
\name{QC-method}
\alias{QC-method}
\alias{QC,AFLP-method}
\title{Extracts the quality control samples of an AFLP object}
\description{
  Extracts the quality control samples of an AFLP object
}
\section{Methods}{
  \describe{

  \item{x = "AFLP"}{Returns the QC information from an AFLP
  object} }
}
\keyword{attribute}
\keyword{methods}

