\docType{methods}
\name{QC<--methods}
\alias{QC<--methods}
\alias{QC<-,AFLP-method}
\title{Adds or overwrites quality control samples}
\description{
  Adds or overwrites quality control samples in an AFLP
  object.
}
\section{Methods}{
  \describe{

  \item{data = "AFLP"}{\code{data} is updated with the new
  quality control samples} }
}
\keyword{methods}

