\docType{methods}
\name{quality-method}
\alias{quality-method}
\alias{quality,AFLP-method}
\title{Extract the quality data from an AFLP object}
\description{
  Extracts the quality from an AFLP object
}
\section{Methods}{
  \describe{

  \item{data = "AFLP"}{Extracts the quality data from an
  AFLP object} }
}
\keyword{attribute}
\keyword{methods}

