\name{quality<--methods}
\docType{methods}
\alias{quality<--method}
\alias{quality<-,AFLP-method}
\title{Adds or overwrites quality information}
\description{
Adds or overwrites quality information in an AFLP object.
}
\section{Methods}{
\describe{

\item{data = "AFLP"}{\code{data} is updated with the new quality information}
}}
\keyword{methods}
