\docType{methods}
\name{replicates-method}
\alias{replicates-method}
\alias{replicates,AFLP-method}
\alias{replicates,AFLP.outlier-method}
\title{Extract replicates from an object}
\description{
  Extract replicates from different kinds of objects
}
\section{Methods}{
  \describe{

  \item{x = "AFLP"}{Returns the vector of replicates in an
  AFLP object}

  \item{x = "AFLP.outlier"}{Returns a data frame with the
  replicates marked as outliers} }
}
\keyword{attribute}
\keyword{methods}

