\docType{methods}
\name{replicates<--methods}
\alias{replicates<--methods}
\alias{replicates<-,AFLP-method}
\title{Adds or overwrites replicates}
\description{
  Adds or overwrites replicates in an AFLP object.
}
\section{Methods}{
  \describe{

  \item{data = "AFLP"}{\code{data} is updated with the new
  replicates} }
}
\keyword{methods}

