\docType{methods}
\name{resid-methods}
\alias{resid-methods}
\alias{resid,AFLP.outlier-method}
\title{Extracts the residual outliers}
\description{
  Extract the residual outliers from an AFLP.outlier object
}
\section{Methods}{
  \describe{ \item{object = "AFLP.outlier"}{returns a
  dataframe with the residual outliers} }
}
\keyword{attribute}
\keyword{methods}

