\name{TiliaDesign}
\docType{data}
\alias{TiliaDesign}
\alias{Tilia_bandvaluespc1}
\alias{Tilia_bandvaluespc2}
\alias{Tilia_bandvaluespc3}
\alias{Tilia_bandvaluespc4}
\title{Raw AFLP data from Tilia samples}
\description{
Several samples were taken from lime trees. The trees were classified in the field as belonging to
either \emph{Tilia cordata}, \emph{Tilia platyphylos} or their hybrid \emph{Tilia europea (x)}. 
}
\usage{TiliaDesign}
\format{A data.frame containing 512 observations.}
     \references{Vander Mijnsbrugge, K and Vanden Broeck, A, unpublished data 
     }
\keyword{datasets}
