\name{AFLP-package}
\Rdversion{1.1}
\alias{AFLP-package}
\alias{AFLP}
\docType{package}
\title{
A package for scoring Amplification Fragment Length Polymorphism (AFLP) data
}
\description{
This packages randomises specimens over different plates while ensuring 
a sufficient number of replicates. The resulting fluorescence data can be
calibrated, scored and analysed.
}
\details{
\tabular{ll}{
Package: \tab AFLP\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2011-03-08\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab methods, lme4, reshape, plyr\cr
Suggests: \tab ggplot2, xtable, vegan\cr
}
}
\author{
Thierry Onkelinx, Paul Quataert

Maintainer: Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}
}
\keyword{ package }
