\docType{methods}
\name{markers-method}
\alias{markers-method}
\alias{markers,AFLP.outlier-method}
\title{Extract the marker information from an object}
\description{
  Extract the marker information from an object
}
\section{Methods}{
  \describe{

  \item{x = "AFLP.outlier"}{Extract the marker information
  from an AFLP.outlier object} }
}
\keyword{attribute}
\keyword{methods}

