\name{outliers}
\Rdversion{1.1}
\alias{outliers}
\title{
Extract outlier information from an AFLP object
}
\description{
Extract outlier information from an AFLP object
}
\usage{
outliers(x)
}
\arguments{
  \item{x}{
	An AFLP object
}
}
\value{
An object of the AFLP.outlier class with the outlier from the AFLP object
}
\author{
Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}, Paul Quataert
}
\seealso{
\code{\link{AFLP-class}}, \code{\link{AFLP.outlier-class}}
}
\examples{
	#data(willowN)
	#outliers(willowN)
}
\keyword{attribute}
