\name{QC-method}
\docType{methods}
\alias{QC-method}
\alias{QC,AFLP-method}
\title{Extract the quality data from an AFLP object}
\description{
Extracts the quality control samples from an AFLP object
}
\section{Methods}{
\describe{

\item{data = "AFLP"}{Extracts the quality control samples from an AFLP object}
}}
\keyword{methods}
\keyword{attribute}
