\name{specimens-methods}
\docType{methods}
\alias{specimens-methods}
\alias{specimens,AFLP-method}
\alias{specimens,AFLP.outlier-method}
\title{Methods for extracting specimen information}
\description{
 Extracts the information on specimen
}
\section{Methods}{
\describe{

\item{x = "AFLP"}{return a vector with the specimen}

\item{x = "AFLP.outlier"}{Return a dataframe with the specimen that are marked as outliers}
}}
\keyword{methods}
\keyword{attribute}
